refactored CmdWithRetries class
moved into a separate helper file
added callbacks for success, fail, ...
added an optional context

+++ main/zb_helpers/zbh_handlers_cmd.hpp
@@ -148,7 +148,7 @@ namespace zb
     inline ZbCmdResponse ZbCmdResponse::g_CmdResponseRegistry[kMaxClusters];
 
 
-    using seq_nr_t = uint8_t;
+    using seq_nr_t = uint16_t;
     /**********************************************************************/
     /* ZbCmdSend                                                          */
     /**********************************************************************/
